extracting memstore compactore out of caompacting memstore
